requirements.txt
# Web-Framework
flask

# MySQL python driver
mysql-connector-python

# Python HTTP library
requests

# HTML parser
selenium
